Update supernova_feedback.h
Stupid mistake

+++ src/physics/supernova_feedback.h
@@ -19,7 +19,7 @@ extern "C"
                                           double new_metals);
   void delayed_supernova_feedback(struct galaxy_t* gal, int snapshot);
   void calc_metal_bubble(struct galaxy_t* gal, int snapshot);
-  double max_array(double arr[30]); 
+  double max_array(float arr[30]); 
   void contemporaneous_supernova_feedback(struct galaxy_t* gal,
                                           double* m_stars,
                                           int snapshot,
